move HasConflicts to shareable location
